requirements.txt
paradex-py>=1.0.0
python-dotenv>=1.0.0
aiohttp>=3.9.0
websockets>=12.0
pyyaml>=6.0
pytest>=7.0
pytest-asyncio>=0.23.0
